Add yatg-trim
yatg-trim is a script to remove ancient data from files generated by
YATG::Store::Disk.

From the perldoc:

This script takes one yatg file, and removes the bottom (i.e. oldest) polls from
that file. YATG itself will read the most recent file in the directory so in theory
the newly generated file will be in use as soon as it is created, and the two
can live side-by-side. In practice you will probably want to delete the old file
after the script has run.